requirements.txt
python-telegram-bot
sqlalchemy
pydantic
starlette
ruamel.yaml
